use hook to filter change (#263)
